asking yes or no to delete

+++ semana7/integrcoes-de-api/labenusers/src/components/ListaUsuarios.js
@@ -21,6 +21,15 @@ const ContainerPagina = styled.div`
   align-items: center;
   justify-content: center;
 `
+const Header = styled.header`
+  height: 60px;
+  background: rgb(9,97,121);
+  background: linear-gradient(166deg, rgba(9,97,121,0.7278207500329179) 11%, rgba(19,69,115,1) 52%, rgba(24,54,111,0.4384949073362886) 76%, rgba(34,36,150,1) 100%);
+  display: flex;
+  align-items: center;
+  padding-left: 18px;
+  margin-bottom: 35px;
+`
 
 const headers = {
     headers:{
@@ -54,18 +63,31 @@ class ListaUsuarios extends React.Component{
   }
 
   //Deletando Usuários
-    deleteUser = (id)=>{
+    deleteUser = (id,name)=>{
         const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`
-        
-        axios.delete(url,headers)
-        .then((resposta)=>{
-            alert("Usuário deletado!")
-            //atualizando lista, sem precisar ficar atualizando a página manualmente
-            this.getAllUsers()
-        })
-        .catch((error)=>{
-            alert("Tente novamente!")
-        })
+        if(window.confirm(`Tem certeza de que deseja deletar, o(a) ${name}?`)){
+          axios.delete(url,headers)
+          .then((resposta)=>{
+              alert("Usuário deletado!")
+              //atualizando lista, sem precisar ficar atualizando a página manualmente
+              this.getAllUsers()
+          })
+          .catch((error)=>{
+              alert("Tente novamente!")
+          })
+        }
+        //--------utilizando: deleteUser = async()=>{} -------------------
+
+        // if(window.confirm("Tem certeza de que deseja deletar?")){
+        //   try{
+        //     const res = await axios.delete(url,headers)
+        //     alert("Usuário deletado!")
+        //     this.getAllUsers()
+  
+        //   }catch (error){
+        //     alert("Tente novamente!")
+        //   }
+        // }   
     }
 
     render(){
@@ -73,15 +95,19 @@ class ListaUsuarios extends React.Component{
         const listaNomesUsuarios = this.state.usuarios.map((cadaUser) =>{
             return <BoxNomes key={cadaUser.id}>
                      {cadaUser.name}
-                     <button onClick={()=>this.deleteUser(cadaUser.id)}>X</button>
+                     <button onClick={()=>this.deleteUser(cadaUser.id, cadaUser.name)}>X</button>
                    </BoxNomes>
         })
         return(
+          <div>
+            <Header>
+            <button onClick={this.props.irParaCadastro}>Página de Cadastro</button>
+            </Header>
             <ContainerPagina>
-                <button onClick={this.props.irParaCadastro}>Página de Cadastro</button>
                 <h1>Lista de Nome dos Usuários Cadastrados</h1>
                 {listaNomesUsuarios}
             </ContainerPagina>
+          </div>
         )
     }
 }
